backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai-agents
openai-chatkit
google-generativeai
qdrant-client
psycopg2-binary
pytest
requests
litellm
sqlalchemy